app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Vladimir Nagibin <vlnagibin@yandex.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.111.0"
pydantic-settings = "2.6.1"
pydantic = "2.10.5"
black = "^24.10.0"
python-multipart = "0.0.20"
sshtunnel = "0.4.0"
SQLAlchemy = "2.0.32"
PyMySQL = "1.1.1"
pandas = "2.2.2"
dropbox = "12.0.2" 
python-dotenv = "1.0.1"
aiofile = "3.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.mypy]
mypy-path = ["./src"]